requirements.txt
pillow
numpy
torch
torchvision 